+++ pkg/mineral/stones.go
@@ -6,6 +6,41 @@ import "github.com/ironarachne/world/pkg/resource"
 func Stones() []Mineral {
 	stones := []Mineral{
 		{
+			Name:         "basalt",
+			PluralName:   "basalt",
+			Hardness:     5,
+			Malleability: 5,
+			Resources: []resource.Resource{
+				{
+					Name:   "basalt",
+					Origin: "basalt",
+					Tags: []string{
+						"stone",
+						"building stone",
+					},
+					Commonality: 5,
+					Value:       1,
+				},
+			},
+		},
+		{
+			Name:         "chalk",
+			PluralName:   "chalk",
+			Hardness:     1,
+			Malleability: 4,
+			Resources: []resource.Resource{
+				{
+					Name:   "chalk",
+					Origin: "chalk",
+					Tags: []string{
+						"stone",
+					},
+					Commonality: 5,
+					Value:       1,
+				},
+			},
+		},
+		{
 			Name:         "granite",
 			PluralName:   "granite",
 			Hardness:     6,
@@ -16,9 +51,10 @@ func Stones() []Mineral {
 					Origin: "granite",
 					Tags: []string{
 						"stone",
+						"building stone",
 					},
 					Commonality: 5,
-					Value:       1,
+					Value:       3,
 				},
 			},
 		},
@@ -33,8 +69,9 @@ func Stones() []Mineral {
 					Origin: "limestone",
 					Tags: []string{
 						"stone",
+						"building stone",
 					},
-					Commonality: 5,
+					Commonality: 10,
 					Value:       1,
 				},
 			},
@@ -50,9 +87,11 @@ func Stones() []Mineral {
 					Origin: "marble",
 					Tags: []string{
 						"stone",
+						"building stone",
+						"sculpture stone",
 					},
 					Commonality: 2,
-					Value:       1,
+					Value:       10,
 				},
 			},
 		},
@@ -65,6 +104,59 @@ func Stones() []Mineral {
 				{
 					Name:   "sandstone",
 					Origin: "sandstone",
+					Tags: []string{
+						"stone",
+						"building stone",
+						"sculpture stone",
+					},
+					Commonality: 10,
+					Value:       1,
+				},
+			},
+		},
+		{
+			Name:         "shale",
+			PluralName:   "shale",
+			Hardness:     1,
+			Malleability: 4,
+			Resources: []resource.Resource{
+				{
+					Name:   "shale",
+					Origin: "shale",
+					Tags: []string{
+						"stone",
+					},
+					Commonality: 5,
+					Value:       1,
+				},
+			},
+		},
+		{
+			Name:         "slate",
+			PluralName:   "slate",
+			Hardness:     4,
+			Malleability: 4,
+			Resources: []resource.Resource{
+				{
+					Name:   "slate",
+					Origin: "slate",
+					Tags: []string{
+						"stone",
+					},
+					Commonality: 5,
+					Value:       1,
+				},
+			},
+		},
+		{
+			Name:         "soapstone",
+			PluralName:   "soapstone",
+			Hardness:     1,
+			Malleability: 4,
+			Resources: []resource.Resource{
+				{
+					Name:   "soapstone",
+					Origin: "soapstone",
 					Tags: []string{
 						"stone",
 					},