+++ pkg/religion/religion.go
@@ -6,11 +6,14 @@ import (
 	"github.com/ironarachne/world/pkg/random"
 )
 
-// Religion is a culture's religion
+// Religion is a religion
 type Religion struct {
+	Name           string
+	CommonName     string
 	Class          Class
 	Pantheon       pantheon.Pantheon
 	GatheringPlace string
+	Virtues        []string
 }
 
 // Generate procedurally generates a religion
@@ -23,6 +26,14 @@ func Generate(originLanguage language.Language) Religion {
 	if religion.Class.PantheonMaxSize > 0 {
 		religion.Pantheon = pantheon.Generate(religion.Class.PantheonMinSize, religion.Class.PantheonMaxSize, originLanguage)
 	}
+
+	name := originLanguage.RandomName()
+	religion.CommonName = name + "ism"
+	name = originLanguage.MakePractice(name)
+	religion.Name = name
+
+	virtues := getRandomVirtues()
+	religion.Virtues = virtues
 
 	return religion
 }